backend/requirements.txt
# Django & DRF
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Redis
django-redis>=5.4.0
redis>=5.0.0

# Async Tasks
celery>=5.3.0
django-celery-beat>=2.6.0
django-celery-results>=2.5.0

# Observability
sentry-sdk[django]>=2.0.0
python-json-logger>=2.0.7

# Real-time (WebSocket)
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0

# Utils
python-dotenv>=1.0.0
gunicorn>=21.2.0
pillow>=10.0.0
cryptography>=42.0.0
bleach>=6.1.0

# API Documentation
drf-spectacular>=0.27.0

# Development
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
freezegun>=1.4.0
